backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
email-validator==2.3.0

# Database dependencies
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.13.1
cryptography==46.0.6